requirements.txt
numpy==1.13.3
Pillow==9.0.0
pycryptodome==3.12.0
